refactor: move initial repos into onboarding route
