+++ src/user-repo/user-repos.module.ts
@@ -2,11 +2,9 @@ import { Module } from "@nestjs/common";
 import { TypeOrmModule } from "@nestjs/typeorm";
 
 import { DbUserRepo } from "./user-repo.entity";
-import { UserReposController } from "./user-repos.controller";
 import { UserReposService } from "./user-repos.service";
 
 @Module({
-  controllers: [UserReposController],
   imports: [
     TypeOrmModule.forFeature([
       DbUserRepo,
